8 модуль таск2
